finished task

+++ app/Http/Controllers/tags.php
@@ -16,7 +16,7 @@ class tags extends Controller
         $request->validate([
             'name'=>['required'],
         ]);
-        $name=$request['tagName'];
+        $name=$request['name'];
         $count=\App\Models\tags::where('tag_name',$name)->count();
         if($count>0)
         {
@@ -38,18 +38,20 @@ class tags extends Controller
         $newName=$request['newName'];
 
 
-        $tag=\App\Models\tags::where('tag_name',$oldName)->first();
-        if($tag->tag_name==$newName)
+
+        if($oldName==$newName)
         {return response('The tag name you entered is the same as the old one.',403);}
         $count=\App\Models\tags::where('tag_name',$newName)->count();
         if($count>0)
         {
-            return response('Tags have to be unique, the tag you enter is duplicate.',403);
+            return response('Tags have to be unique, the tag you entered is duplicate.',403);
         }
+        $tag=\App\Models\tags::where('tag_name',$oldName)->first();
+
         $tag->tag_name=$newName;
         $tag->save();
 
-        return response('Tag '.$oldName.' updated successfully.',200);
+        return response($oldName.' tag updated successfully.',200);
     }
     public function deleteTag(Request $request)
     {
@@ -58,8 +60,10 @@ class tags extends Controller
         ]);
         $name=$request['name'];
         $tag=\App\Models\tags::where('tag_name',$name)->first();
+        if ($tag==null)
+            return response('Tag not found, please make sure that you entered the right name.');
         $tag->delete();
 
-        return response('Tag '.$name.' deleted successfully.',200);
+        return response($name.' tag deleted successfully.',200);
     }
 }
